requirements.txt
aniso8601==8.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
future==0.18.2
idna==2.10
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
PyJWT==1.7.1
pyparsing==2.4.7
pyserial
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
twilio==6.48.0
urllib3==1.26.2
Werkzeug==1.0.1